Replace SingletonMeta by simple static module
SingletonMeta had a very hard to fix deadlock: When a class inheriting from SingletonMeta tried to construct a different SingletonMeta in its constructor, the second SingletonMeta instance could never be constructed, because the check for the class happened only AFTER trying to acquire the lock, instead of before.

It turns out that in the case of the ROSBridge, an explicit Singleton metaclass wasn't
even necessary. So I removed the metaclass and switched ROSBridge to a
simple static module, whose globals will only be instantiated once by
the Python import system.